moved client/server code to corresponding files
